requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
pydantic==1.10.11
httpx==0.24.1
pytest==7.4.0
python-dotenv==1.0.0